requirements.txt
asyncpg==0.30.0
cryptography==46.0.2
email_validator==2.2.0
fastapi-utils==0.7.0
greenlet==3.2.4
h2==4.3.0
pip==25.2
pydantic-settings==2.10.1
python-jose==3.3.0
slowapi==0.1.9
sqlmodel==0.0.24
supabase==2.3.4
typing-inspect==0.9.0
uvicorn==0.34.3
